// src/app/model/Geometry.ts
/**
 * API documentation
 * To use this api you need an access token. To get the token, send a post request with your email address to api-users resource and one will be send to your. See below for information on how to use this api and if you have any questions you can contact us at helpdesk@laji.fi.  Place refer to [schema.laji.fi](http://schema.laji.fi/) for more information about the used vocabulary
 *
 * OpenAPI spec version: 0.1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * GeoJSon geometry
 */
export interface Geometry {
    /**
     * the geometry type
     */
    type: Geometry.TypeEnum;

    coordinates: any;

}
export namespace Geometry {
    export enum TypeEnum {
        Point = <any> 'Point',
        LineString = <any> 'LineString',
        Polygon = <any> 'Polygon',
        MultiPoint = <any> 'MultiPoint',
        MultiLineString = <any> 'MultiLineString',
        MultiPolygon = <any> 'MultiPolygon'
    }
}